requirements.txt
flask==3.0.3
openai>=1.40.0
markdown==3.6
python-dotenv==1.0.1